shop/requirements.txt
Django>=5.2.0,<6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
pycryptodome>=3.19.0
requests>=2.31.0 